% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFam.R
\name{check_OCFam_args}
\alias{check_OCFam_args}
\title{Internal: basic checks on scalar arguments for OCFam()}
\usage{
check_OCFam_args(
  kinship_constraint,
  iterations,
  sire_count_by_age,
  dam_count_by_age,
  min_prop_fams,
  max_parents_per_fam
)
}
\description{
Internal: basic checks on scalar arguments for OCFam()
}
