% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFam.R
\name{check_OCFam_ped}
\alias{check_OCFam_ped}
\title{Internal: basic checks on 'ped' for OCFam()}
\usage{
check_OCFam_ped(ped)
}
\description{
Ensures that the pedigree has all required columns and basic
structure/typing for OCFam to work safely.
}
