% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFamPrep.R
\name{check_OCFamPrep_inputs}
\alias{check_OCFamPrep_inputs}
\title{Internal: basic checks on inputs for OCFamPrep()}
\usage{
check_OCFamPrep_inputs(ped, age_class_names, gene_flow_vector)
}
\description{
Ensures that the input pedigree and optional age-class and gene-flow
information are structurally consistent with what OCFamPrep expects.
}
\keyword{internal}
