% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiSelFam.R
\name{optiSelFam}
\alias{optiSelFam}
\title{optiSelFam}
\usage{
optiSelFam(
  candidate_parents,
  ped,
  indiv_contbn,
  kinship_constraint,
  step_interval = 0.1,
  overlapping_gens,
  gene_flow_vector,
  min_prop_parent_fams_to_retain_in_youngest_age_class
)
}
\arguments{
\item{candidate_parents}{is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'Indiv' (character).}
 \item{'Contbn_count'  (NA if unknown) (numeric). Can Contbn_count and Exclude_max_parents_per_fam be the one column e.g. 0 = fixed contbn NA = no constraint??????? - FIXED_CONTBN}
 \item{'lb'  (numeric). These would be more logically defined as counts - MIN_CONTBN (integers)??????}
 \item{'ub'  (numeric). These would be more logically defined as counts - MAX_CONTBN (integers)??????}
 \item{'Exclude_max_parents_per_fam'  (logical). Could just call this EXCLUDE??????}
}}

\item{ped}{is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'Indiv' (character).}
 \item{'Sire'  (character).}
 \item{'Dam'  (character).}
 \item{'Fam'  (character).}
 \item{'Born'  (numeric).               is this required - could obtain it from ped????}
 \item{'EBV'  (numeric).}
 \item{'RANK'  (integer).               is this required - could obtain it from EBV????}
 \item{'FAM_SIRE'  (character).        is this required - could obtain it from ped????}
 \item{'FAM_DAM'  (character).         is this required - could obtain it from ped????}
 \item{'LINE'  (character).          is this required????}
 \item{'COHORT'  (character).        is this required????}
 \item{'AVAIL_BROOD'  (logical).       is this required - could obtain it from candidate_parents????}
 \item{'Breed'  (character).         is this required????}
 \item{'Contbn_count'  (integer).      is this required - could obtain it from candidate_parents????}
 \item{'AVAIL_OR_PAST_BROOD'  (logical).  is this required - could obtain it from candidate_parents????}
}}

\item{indiv_contbn}{is the ...... (numeric between 0 and 1)}

\item{kinship_constraint}{is the ...... (numeric between 0 and 1)}

\item{step_interval}{is the ...... (numeric between 0 and 1)???}

\item{gene_flow_vector}{is a vector ......  as.numeric(NA) if overlapping_gens=FALSE (numeric). For example, gene_flow_vector = c(0.2, 0.8, 0, 0) - Year 4, 3, 2, 1, oldest to youngest}

\item{min_prop_parent_fams_to_retain_in_youngest_age_class}{is the ...... (numeric between 0 and 1)}
}
\value{
'fam_contbn' is a data frame containing details of contributions by family:
\itemize{
 \item{ }
}

'parent_contbn' is a data frame containing details of contributions by individual:
\itemize{
 \item{}
}

'fit_out' is a vector containing details of the constraints applied:
\itemize{
 \item{}
}
}
\description{
This function generates a mating list for a set of parents.
The mating list can be generated i) to minimise the average inbreeding coefficient (F) of families generated or ii) according to assortative mating principles.
Inputs include a list of parents and a 3-column pedigree file specifying the ancestry of these candidates.
}
\examples{
#Retrieve example data
candidate_parents <- optiSelFam::candidate_parents
ped <- optiSelFam::ped

output <- optiSelFam(
candidate_parents = candidate_parents,
ped = ped,
indiv_contbn = (1/180),
kinship_constraint = 0.09,
step_interval = 0.1,
overlapping_gens = FALSE,
gene_flow_vector = NA,
min_prop_parent_fams_to_retain_in_youngest_age_class = 1)
head(output$fam_contbn)
head(output$parent_contbn)
output$fit_out
}
