% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFam.R, R/OCFam_02.R
\name{OCFam}
\alias{OCFam}
\title{OCFam}
\usage{
OCFam(
  ped,
  N_fams,
  kinship_constraint,
  step_interval = 0.1,
  overlapping_gens,
  gene_flow_vector,
  min_prop_fams,
  max_parents_per_fam
)

OCFam(
  ped,
  N_fams,
  kinship_constraint,
  step_interval = 0.1,
  overlapping_gens,
  gene_flow_vector,
  min_prop_fams,
  max_parents_per_fam
)
}
\arguments{
\item{ped}{is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'INDIV' (character).}
 \item{'SIRE'  (character).}
 \item{'DAM'  (character).}
 \item{'FAM'  (character).}
 \item{'BORN'  (numeric).}
 \item{'EBV'  (numeric).}
 \item{'N_AS_PARENT_CURRENT'  (numeric).} #number of families contributed to in the current spawn round
 \item{'AVAIL_BROOD'  (logical).}
}}

\item{kinship_constraint}{is the ...... (numeric between 0 and 1)}

\item{step_interval}{is the ...... (numeric between 0 and 1)???}

\item{gene_flow_vector}{is a vector ......  as.numeric(NA) if overlapping_gens=FALSE (numeric). For example, gene_flow_vector = c(0.2, 0.8, 0, 0) - Year 4, 3, 2, 1, oldest to youngest}

\item{min_prop_fams}{is the min_prop_parent_fams_to_retain_in_youngest_age_class (numeric between 0 and 1)}

\item{max_parents_per_fam}{is the ...... (integer)}

\item{indiv_contbn}{is the ...... (numeric between 0 and 1)}
}
\value{
'fam_contbn' is a data frame containing details of contributions by family:
\itemize{
 \item{ }
}

'parent_contbn' is a data frame containing details of contributions by individual:
\itemize{
 \item{}
}

'fit_out' is a vector containing details of the constraints applied:
\itemize{
 \item{}
}

'fam_contbn' is a data frame containing details of contributions by family:
\itemize{
 \item{ }
}

'parent_contbn' is a data frame containing details of contributions by individual:
\itemize{
 \item{}
}

'fit_out' is a vector containing details of the constraints applied:
\itemize{
 \item{}
}
}
\description{
This function generates a mating list for a set of parents.
The mating list can be generated i) to minimise the average inbreeding coefficient (F) of families generated or ii) according to assortative mating principles.
Inputs include a list of parents and a 3-column pedigree file specifying the ancestry of these candidates.

This function generates a mating list for a set of parents.
The mating list can be generated i) to minimise the average inbreeding coefficient (F) of families generated or ii) according to assortative mating principles.
Inputs include a list of parents and a 3-column pedigree file specifying the ancestry of these candidates.
}
\examples{
#Retrieve example data
candidate_parents <- OCFam::candidate_parents
ped <- OCFam::ped

output <- OCFam(
candidate_parents = candidate_parents,
ped = ped,
indiv_contbn = (1/180),
kinship_constraint = 0.09,
step_interval = 0.1,
overlapping_gens = FALSE,
gene_flow_vector = NA,
min_prop_fams = 1)
head(output$fam_contbn)
head(output$parent_contbn)
output$fit_out
#Retrieve example data
candidate_parents <- OCFam::candidate_parents
ped <- OCFam::ped

output <- OCFam(
candidate_parents = candidate_parents,
ped = ped,
indiv_contbn = (1/180),
kinship_constraint = 0.09,
step_interval = 0.1,
overlapping_gens = FALSE,
gene_flow_vector = NA,
min_prop_fams = 1)
head(output$fam_contbn)
head(output$parent_contbn)
output$fit_out
}
