% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFam.R, R/OCFam_02.R
\name{OCFam}
\alias{OCFam}
\title{OCFam}
\usage{
OCFam(
  ped,
  N_fams,
  kinship_constraint,
  step_interval = 0.1,
  gene_flow_vector,
  min_prop_fams,
  max_parents_per_fam
)

OCFam(
  ped,
  N_fams,
  kinship_constraint,
  step_interval = 0.1,
  gene_flow_vector,
  min_prop_fams,
  max_parents_per_fam
)
}
\arguments{
\item{ped}{is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'Indiv' (character).}
 \item{'Sire'  (character).}
 \item{'Dam'  (character).}
 \item{'Fam'  (character).}
 \item{'Born'  (numeric).}
 \item{'EBV'  (numeric).}
 \item{'N_AS_PARENT_CURRENT'  (numeric).} #number of families contributed to in the current spawn round
 \item{'AVAIL_BROOD'  (logical).}
}}

\item{N_fams}{is the number of families to be generated in the next age class including those already produced (see N_AS_PARENT_CURRENT in ped) (integer)}

\item{kinship_constraint}{is the maximum value of the mean kinship between families in the next age class, while mean EBV is maximised (max.EBV).  If NA there is no constraint placed on kinship and the average kinship is minimised while EBV is not considered (min.fPED) (numeric between 0 and 1)}

\item{step_interval}{is a parameter that controls rounding error in individual contributions.  The larger the value the greater the rounding error but the quicker it runs (numeric between 0 and 1)}

\item{gene_flow_vector}{is a applicable to overlapping generations (if NA discrete generations is assumed).  It is vector representing parental contributions by age class to the next age class. For example, gene_flow_vector = c(0.2, 0.8, 0, 0) - oldest age class to youngest age class.}

\item{min_prop_fams}{is the proportion of families to be retained (i.e. to contribute at least one parent) from the oldest age class with parental candidates (numeric between 0 and 1)}

\item{max_parents_per_fam}{is the maximum number of parents to contribute from each family (integer)}

\item{indiv_contbn}{is the ...... (numeric between 0 and 1)}
}
\value{
'fam_contbn' is a data frame containing details of contributions by family:
\itemize{
 \item{ }
}

'parent_contbn' is a data frame containing details of contributions by individual:
\itemize{
 \item{}
}

'fit_out' is a vector containing details of the constraints applied:
\itemize{
 \item{}
}

'fam_contbn' is a data frame containing details of contributions by family:
\itemize{
 \item{FAM}
 \item{N_INDIV_TOTAL}
 \item{N_AS_PAST_PARENT}
 \item{N_INDIV}
}

'parent_contbn' is a data frame containing details of contributions by individual:
\itemize{
 \item{INDIV}
 \item{SIRE}
 \item{DAM}
 \item{FAM}
 \item{BORN    }
 \item{EBV}
 \item{AVAIL_BROOD}
 \item{N_AS_PARENT_CURRENT}
 \item{RANK}
 \item{AVAIL_OR_PAST_BROOD }
 \item{"FAM_SIRE}
 \item{FAM_DAM}
 \item{BREED}
 \item{MATURE}
 \item{IS_CANDIDATE}
 \item{N_TOTAL_AS_PARENT}
 \item{ADDED_IN_LAST_ITERATION}
}

'candidate_parents' is a data frame containing details of candidate parents:
\itemize{
 \item{INDIV}
 \item{N_AS_PARENT_CURRENT}
 \item{LB}
 \item{UB}
 \item{EXCLUDE_MAX_PARENTS_PER_FAM}
 \item{BORN}
}

'fit_out' is a vector containing details of the constraints applied in the implementation of optiSel:
}
\description{
This function generates a mating list for a set of parents.
The mating list can be generated i) to minimise the average inbreeding coefficient (F) of families generated or ii) according to assortative mating principles.
Inputs include a list of parents and a 3-column pedigree file specifying the ancestry of these candidates.

This function generates a mating list for a set of parents.
The mating list can be generated i) to minimise the average inbreeding coefficient (F) of families generated or ii) according to assortative mating principles.
Inputs include a list of parents and a 3-column pedigree file specifying the ancestry of these candidates.
}
\examples{
#Retrieve example data
candidate_parents <- OCFam::candidate_parents
ped <- OCFam::ped

output <- OCFam(
candidate_parents = candidate_parents,
ped = ped,
indiv_contbn = (1/180),
kinship_constraint = 0.09,
step_interval = 0.1,
overlapping_gens = FALSE,
gene_flow_vector = NA,
min_prop_fams = 1)
head(output$fam_contbn)
head(output$parent_contbn)
output$fit_out
#Retrieve example data
candidate_parents <- OCFam::candidate_parents
ped <- OCFam::ped

output <- OCFam(
candidate_parents = candidate_parents,
ped = ped,
indiv_contbn = (1/180),
kinship_constraint = 0.09,
step_interval = 0.1,
gene_flow_vector = NA,
min_prop_fams = 1)
head(output$fam_contbn)
head(output$parent_contbn)
output$fit_out
}
