% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFam.R
\name{OCFam}
\alias{OCFam}
\title{OCFam}
\usage{
OCFam(
  ped,
  N_fams,
  kinship_constraint,
  step_interval = 0.1,
  gene_flow_vector,
  min_prop_fams,
  max_parents_per_fam
)
}
\arguments{
\item{ped}{is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'INDIV' is the individual identifier (character).}
 \item{'SIRE' is the male parent identifier (character).}
 \item{'DAM' is the female parent identifier (character).}
 \item{'FAM' is a full-sibling family identifier (character).}
 \item{'BORN' integer indicating age class.  May be the year of birth if one age class per year or an integer indicating the sequence of age classes (integer).}
 \item{'EBV' is the estimated breeding value (numeric).}
 \item{'N_AS_PARENT_PREV' is the number of families in the next age class previously contributed to (integer).}
 \item{'AVAIL_BROOD' is TRUE if the individual is candidate parent of next age class (logical).}
 }}

\item{N_fams}{is the number of families to be generated in the next age class including those already produced (see N_AS_PARENT_PREV in ped) (integer)}

\item{kinship_constraint}{is the maximum value of the mean kinship between families in the next age class, while mean EBV is maximised (max.EBV).  If NA there is no constraint placed on kinship and the average kinship is minimised while EBV is not considered (min.fPED) (numeric between 0 and 1)}

\item{step_interval}{is a parameter that controls rounding error in individual contributions.  The larger the value the greater the rounding error but the quicker it runs (numeric between 0 and 1)}

\item{gene_flow_vector}{is a applicable to overlapping generations (if NA discrete generations is assumed).  It is vector representing parental contributions by age class to the next age class. For example, gene_flow_vector = c(0.2, 0.8, 0, 0) - oldest age class to youngest age class.}

\item{min_prop_fams}{is the proportion of families to be retained (i.e. to contribute at least one parent) from the oldest age class with parental candidates (numeric between 0 and 1)}

\item{max_parents_per_fam}{is the maximum number of parents to contribute from each family (integer)}
}
\value{
'fam_contbn' is a data frame containing details of contributions by family:
\itemize{
 \item{FAM}
 \item{N_INDIV_TOTAL}
 \item{N_AS_PAST_PARENT}
 \item{N_INDIV}
}

'parent_contbn' is a data frame containing details of contributions by individual:
\itemize{
 \item{'INDIV' is the individual identifier (character).}
 \item{'SIRE' is the male parent identifier (character).}
 \item{'DAM' is the female parent identifier (character).}
 \item{'FAM' is a full-sibling family identifier (character).}
 \item{'BORN' integer indicating age class.  May be the year of birth if one age class per year or an integer indicating the sequence of age classes (integer).}
 \item{'EBV' is the estimated breeding value (numeric).}
 \item{'N_AS_PARENT_PREV' is the number of families in the next age class previously contributed to (integer).}
 \item{'AVAIL_BROOD' is TRUE if the individual is candidate parent of next age class (logical).}
 \item{'RANK' individual rank based on EBV}
 \item{'AVAIL_OR_PAST_BROOD' is TRUE if AVAIL_BROOD is TRUE or the individual has previously been used as a parent in the next age class.}
 \item{'FAM_SIRE' is a full-sibling family identifier of the SIRE (character).}
 \item{'FAM_DAM' is a full-sibling family identifier of the DAM (character).}
 \item{'BREED' is the species.}
 \item{'MATURE' is sexually mature}
 \item{'IS_CANDIDATE' was considered a candidate parent in OCFam}
 \item{'N_TOTAL_AS_PARENT' is the total count as a parent in the next year class}
 \item{'ADDED_IN_LAST_ITERATION' individual was addes in the last iteration of OCFam.  Would be the first individuals to be removed in necessary.}
}

'candidate_parents' is a data frame containing details of candidate parents:
\itemize{
 \item{'INDIV' is the individual identifier (character).}
 \item{'N_AS_PARENT_PREV' is the number of families contributed to in the next age class prior to running optimal contributions (integer).}
 \item{'LB' is the low bound of possilbe contribution to the next age class}
 \item{'UB' is the upper bound of possilbe contribution to the next age class}
 \item{EXCLUDE_MAX_PARENTS_PER_FAM' If TRUE individual is excluded on the basis that there are better parents available from the family to meet the max_parents_per_fam constraint.}
}

'fit_out' is a vector containing details of the constraints applied in the implementation of optiSel:
}
\description{
This function implements the optimal contributions method of Hamilton (2020) https://doi.org/10.1093/jhered/esaa051
It also addresses the rounding issue associated with standard optimal contributions, particularly in highly fecund species in which relatively few families are generated (or parents are used).
}
\examples{
#Retrieve example data
ped <- OCFam::ped
tail(ped)

#Run OCFam function
OCFam_output <- OCFam::OCFam(ped = ped,
                             N_fams = 60,
                             kinship_constraint = 0.012,
                             step_interval = 0.1,
                             gene_flow_vector = NA,
                             min_prop_fams = 0.9,
                             max_parents_per_fam = 4
)

OCFam_output$fit_out$summary
OCFam_output$fit_out$mean
head(OCFam_output$fam_contbn)
}
