% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCFamPrep.R
\name{OCFamPrep}
\alias{OCFamPrep}
\title{OCFamPrep}
\usage{
OCFamPrep(ped, age_class_names = NULL, gene_flow_vector = NULL)
}
\arguments{
\item{ped}{is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'INDIV' is the individual identifier (character).}
 \item{'SIRE' is the male parent identifier (character). Unknown sires should be entered as '0' or NA.}
 \item{'DAM' is the female parent identifier (character). Unknown sires should be entered as '0' or NA.}
 \item{'FAM' is a full-sibling family identifier.  May be NA if parents are unknown (character).}
 \item{'BORN' integer indicating age class.  May be the year of birth if one age class per year or an integer indicating the sequence of age classes (integer).}
 \item{'EBV' is the estimated breeding value. This field may contain missing values or NA (numeric).}
 }}

\item{age_class_names}{is a data frame specifying names for age classes (class in parentheses):
\itemize{
 \item{'AGE_CLASS_NAME' is a name for the age class (character).}
 \item{'BORN' integer indicating age class.  May be the year of birth if one age class per year or an integer indicating the sequence of age classes (integer).}
 }}

\item{gene_flow_vector}{is a applicable to overlapping generations (if NA discrete generations is assumed).  It is vector representing the assumed parental contributions by age class to the next age class. For example, gene_flow_vector = c(0.2, 0.8, 0, 0) assumes 20 percent of parents are four age classes old and 80 percent are three age classes old.}
}
\value{
'plot' is plot of trends in coancestry and inbreeding coefficients

'age_class_means' is a dataframe listing inbreeding coefficients for individuals within families.
\itemize{
 \item{'BORN' is age class (integer).}
 \item{'INBREEDING' is the average Wright's inbreeding coefficient of families in the age class (numeric).}
 \item{'COANCESTRY' is the average coancestry between families in the age class (numeric).}
 \item{'OVERLAPPING_COANCESTRY' is the average coancestry between families weighted by the age class contributions specified in the gene_flow_vector (numeric).}
}

'age_class_K_mat' is a matrix containing average coancestries within age classes on the diagonals and between age classes on the off-diagonals.  Row and column names are the FAM identifiers from 'ped'.

'fam_K_matrix' is a data frame containing the between family coancestry matrix (half of the Between-Family Relationship Matrix described in Hamilton (2020) 'Optimal Contribution Selection in Highly Fecund Species With Overlapping Generations').  Row and column names are the FAM identifiers from 'ped'.

'family_inbreeding' is a dataframe listing inbreeding coefficients for individuals within families.'
\itemize{
 \item{'FAM' is the family identifier (character).}
 \item{'F' is the Wright's inbreeding coefficient for individuals in the family (numeric).}
}

'ped' is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'INDIV' is the individual identifier (character).}
 \item{'SIRE' is the male parent identifier (character).}
 \item{'DAM' is the female parent identifier (character).}
 \item{'FAM' is a full-sibling family identifier (character).}
 \item{'BORN' integer indicating age class.  May be the year of birth if one age class per year or an integer indicating the sequence of age classes (integer).}
 \item{'EBV' is the estimated breeding value (numeric).}
 \item{'FAM_SIRE' is a full-sibling family identifier of the SIRE (character).}
 \item{'EBV_SIRE' is the estimated breeding value of the SIRE (numeric).}
 \item{'FAM_DAM' is a full-sibling family identifier of the DAM (character).}
 \item{'EBV_DAM'  is the estimated breeding value of the DAM (numeric).}
 \item{'EBV_FAM' is the mean EBV of the FAM's SIRE and DAM}
 }

'family_ped' is a data frame with the following columns (class in parentheses):
\itemize{
 \item{'FAM' is a full-sibling family identifier (character).}
 \item{'SIRE' is the male parent identifier (character).}
 \item{'DAM' is the female parent identifier (character).}
 \item{'BORN' integer indicating age class.  May be the year of birth if one age class per year or an integer indicating the sequence of age classes (integer).}
 \item{'FAM_SIRE' is a full-sibling family identifier of the SIRE (character).}
 \item{'EBV_SIRE' is the estimated breeding value of the SIRE (numeric).}
 \item{'FAM_DAM' is a full-sibling family identifier of the DAM (character).}
 \item{'EBV_DAM'  is the estimated breeding value of the DAM (numeric).}
 \item{'EBV_FAM' is the mean EBV of the FAM's SIRE and DAM}
 }

'gene_flow_vector' is a applicable to overlapping generations (if NA discrete generations is assumed).  It is vector representing the assumed parental contributions by age class to the next age class. For example, gene_flow_vector = c(0.2, 0.8, 0, 0) assumes 20 percent of parents are four age classes old and 80 percent are three age classes old.

'r' is a weight vector of age classes; r[i] denotes the long-term contribution of age class i (until lifetime is reached).  See Meuwissen, T. H. E. and A. K. Sonesson. 1998. Maximizing the response of selection with a predefined rate of inbreeding: overlapping generations.

'gen_interval' is the generation interval.
}
\description{
This function provides tables and a plot that allows analysis of historic trends in average coancestry (k) and inbreeding (Wrights inbreeding coefficient, F) across year classes.
}
\examples{
#Retrieve example data
ped_small <- OCFam::ped_small
age_class_names <-  data.frame(AGE_CLASS_NAME = c("Founders", "Gen_0", "Gen_1", "Gen_2"), BORN = c(0, 1, 2, 3))
gene_flow_vector <- c(0.2,0.8)

#Show example data
tail(ped_small)
age_class_names
gene_flow_vector

#Run OCFam function
OCFamPrep_output <- OCFam::OCFamPrep(ped = ped_small,
                             age_class_names = age_class_names,
                             gene_flow_vector = gene_flow_vector
)

OCFamPrep_output$age_class_means
OCFamPrep_output$age_class_K_mat
OCFamPrep_output$fam_K_matrix[1:5,1:5]
head(OCFamPrep_output$family_inbreeding)

}
